pyproject.toml
[tool.poetry]
name = "advanced-cloud-deployment"
version = "0.1.0"
description = "Advanced Cloud Deployment - Todo Chatbot with Kafka and Dapr"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
sqlmodel = "^0.0.16"
pydantic = "^2.5.0"
httpx = "^0.25.2"
python-multipart = "^0.0.6"
python-dateutil = "^2.8.2"
aiokafka = "^0.10.0"
kafka-python = "^2.0.2"
dapr-ext-fastapi = "^1.12.0"
apscheduler = "^3.10.4"
asyncio = "^3.4.3"
pytz = "^2023.3.post1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "api.main:main"
dev = "api.main:dev"